refactor: hook 분리 및 search input 분리 - 복잡도 관리
